move symbols of psyke/regression/{feature_not_found_exception, hypercube}.py into psyke/regression/__init__.py
