--- psyke/regression/feature_not_found_exception.py
@@ -1,4 +1,0 @@
-class FeatureNotFoundException(Exception):
-
-    def __init__(self, feature: str):
-        super().__init__('Feature "' + feature + '" not found.')
